byte_infer_perf/general_perf/backends/ILUVATAR/optimizer/requirements.txt
sympy
packaging
onnxsim